Create README.md
hibernate-when-idle
==============

This is a shell script to hibernate a Linux system after a configurable idle period.

It is different to other alternatives in that it is intended no only within an X session. It should monitor activity whether you're on an active session, a terminal session, or in the session manager screen (kdm, gdm, etc). It is a work in progress (incomplete and buggy).